[rfc]Load more

+++ imports/api/posts/server/securedMethods.js
@@ -3,19 +3,19 @@ import PostService from '/imports/api/posts/postService';
 
 Meteor.methods({
     'secured.post_create'(post) {
-        return PostService.createPost(post);
+        return PostService.createPost({...post, userId: this.userId});
     },
 
     'secured.post_list' () {
-        return PostService.listPosts();
+        return PostService.listPosts({userId: this.userId});
     },
 
-    'secured.post_edit' (_id, postData) {
-        PostService.editPost(_id, postData);
+    'secured.post_edit' ({_id, post}) {
+        PostService.editPost({_id, userId: this.userId, post});
     },
 
     'secured.post_remove' (_id){
-        PostService.removePost(_id);
+        PostService.removePost({_id, userId: this.userId});
     },
 
     'secured.post_get' (_id) {
